Move response parsing to uploader
