+++ packages/uploader/lib/src/http/session/network_delegate.dart
@@ -1,19 +1,15 @@
-import 'dart:convert';
 import 'package:cloudinary_dart/src/http/request/multi_part_request.dart';
-import 'package:cloudinary_dart/uploader/uploader_response.dart';
 import 'package:cloudinary_dart_url_gen/cloudinary.dart';
-import 'package:http/http.dart';
 import 'package:http/http.dart' as http;
+import 'package:http/http.dart';
 
 import '../../request/network_request.dart';
-import '../../response/upload_error.dart';
-import '../../response/upload_result.dart';
 
 class NetworkDelegate {
 
   NetworkDelegate();
 
-  Future<UploaderResponse<UploadResult>> callApi(NetworkRequest request) async {
+  Future<StreamedResponse> callApi(NetworkRequest request) async {
     StreamedResponse requestResponse;
     request.headers.addEntries(addUserAgent());
     var multiPartRequest =
@@ -22,34 +18,7 @@ class NetworkDelegate {
     multiPartRequest.files
         .add(await http.MultipartFile.fromPath('file', request.payload.path));
     requestResponse = await multiPartRequest.send();
-    return processResponse(requestResponse, request.adapter);
-  }
-
-  Future<UploaderResponse<UploadResult>> processResponse(
-      http.StreamedResponse? requestResponse, String adapter) {
-    return getProcessedResponse(
-        requestResponse!.statusCode,
-        requestResponse!.stream,
-        requestResponse!.headers['x-cld-error'],
-        adapter);
-  }
-
-  Future<UploaderResponse<UploadResult>> getProcessedResponse(int statusCode,
-      ByteStream stream, String? errorHeader, String adapter) async {
-    String body = await stream.transform(utf8.decoder).join();
-    if (statusCode >= 200 && statusCode <= 299) {
-      if (body != null) {
-        var response =
-            UploaderResponse<UploadResult>(statusCode, body, null, body);
-        return response;
-      } else {
-        var responseError = UploadError("Error");
-        return UploaderResponse(statusCode, null, responseError, body);
-      }
-    } else {
-      return UploaderResponse(
-          statusCode, null, UploadError(errorHeader ?? "Unknown Error"), body);
-    }
+    return requestResponse;
   }
 
   Map<String, String> paramsToFields(Map<String, dynamic> params) {
