+++ packages/uploader/lib/uploader/uploader.dart
@@ -1,9 +1,12 @@
+import 'dart:convert';
 import 'dart:io';
 import 'package:cloudinary_dart/src/http/extensions/list_extension.dart';
 import 'package:cloudinary_dart/uploader/abstract_uploader_request.dart';
 import 'package:cloudinary_dart/uploader/uploader_response.dart';
 import 'package:cloudinary_dart_url_gen/cloudinary.dart';
 import 'package:cloudinary_dart_url_gen/config/api_config.dart';
+import 'package:http/http.dart';
+
 
 import '../src/http/session/network_delegate.dart';
 import '../src/http/request/multi_part_request.dart';
@@ -11,6 +14,7 @@ import '../src/request/network_request.dart';
 import '../src/request/params/upload_params.dart';
 import '../src/request/payload.dart';
 import '../src/request/upload_request.dart';
+import '../src/response/upload_error.dart';
 import '../src/response/upload_result.dart';
 
 class Uploader {
@@ -22,8 +26,8 @@ class Uploader {
 
 
   // Api calls
-  Future<UploaderResponse<UploadResult>?> upload(File file, {UploadParams? params, ProgressCallback? progressCallback, Map<String, String>? extraHeaders}) {
-    FilePayload payload = FilePayload(file);
+  Future<UploaderResponse<UploadResult>?> upload(dynamic file, {UploadParams? params, ProgressCallback? progressCallback, Map<String, String>? extraHeaders}) {
+    Payload<dynamic> payload = buildPayload(file);
     UploadRequest request = UploadRequest(this, params ?? UploadParams(), payload, progressCallback: progressCallback, extraHeaders: extraHeaders);
     return request.execute();
   }
@@ -54,7 +58,7 @@ class Uploader {
   }
 
   Future<UploaderResponse<UploadResult>> callApi(AbstractUploaderRequest request, String action, String adapter) async {
-    return await networkDelegate.callApi(prepareNetworkRequest(action, (request as UploadRequest), adapter));
+    return processResponse(await networkDelegate.callApi(prepareNetworkRequest(action, (request as UploadRequest), adapter)));
   }
 
   Future<UploaderResponse<UploadResult>> performUpload(UploadRequest request) {
@@ -75,5 +79,42 @@ class Uploader {
 
   bool isRemoteUrl(String value) {
     return RegExp('ftp:.*|https?:.*|s3:.*|gs:.*|data:([w-]+/[w-]+)?(;[w-]+=[w-]+)*;base64,([a-zA-Z0-9/+ =]+)').hasMatch(value);
+  }
+
+  Payload buildPayload(dynamic file) {
+    if (file is File) {
+      return FilePayload(file);
+    } else if(file is String) {
+      return UrlPayload(file);
+    } else if(file is Stream) {
+      return StreamPayload(file);
+    }
+    throw ArgumentError("Current file type is not supported");
+  }
+
+  Future<UploaderResponse<UploadResult>> processResponse(
+      StreamedResponse? requestResponse) {
+    return getProcessedResponse(
+        requestResponse!.statusCode,
+        requestResponse!.stream,
+        requestResponse!.headers['x-cld-error']);
+  }
+
+  Future<UploaderResponse<UploadResult>> getProcessedResponse(int statusCode,
+      ByteStream stream, String? errorHeader) async {
+    String body = await stream.transform(utf8.decoder).join();
+    if (statusCode >= 200 && statusCode <= 299) {
+      if (body != null) {
+        var response =
+        UploaderResponse<UploadResult>(statusCode, body, null, body);
+        return response;
+      } else {
+        var responseError = UploadError("Error");
+        return UploaderResponse(statusCode, null, responseError, body);
+      }
+    } else {
+      return UploaderResponse(
+          statusCode, null, UploadError(errorHeader ?? "Unknown Error"), body);
+    }
   }
 }
